Use pos-visible-in-window-p.
Unlike the old code, this method for detecting off-window positions
also works correctly when the bottom-most line of the window is only
partially shown.

With the old code, moving point into the partially shown line would not
trigger a smooth scroll in sublimity--post-command, but trigger a
standard Emacs scroll later in redisplay. With the new code, moving
point into the partially shown line triggers a smooth scroll.